+++ src/Components/Component.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Templateless\Components;
+
+interface Component {}