+++ src/Components/Service.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Templateless\Components;
+
+enum Service: string
+{
+    case WEBSITE = 'WEBSITE';
+    case EMAIL = 'EMAIL';
+    case PHONE = 'PHONE';
+    case FACEBOOK = 'FACEBOOK';
+    case YOUTUBE = 'YOUTUBE';
+    case TWITTER = 'TWITTER';
+    case X = 'X';
+    case GITHUB = 'GITHUB';
+    case INSTAGRAM = 'INSTAGRAM';
+    case LINKEDIN = 'LINKEDIN';
+    case SLACK = 'SLACK';
+    case DISCORD = 'DISCORD';
+    case TIKTOK = 'TIKTOK';
+    case SNAPCHAT = 'SNAPCHAT';
+    case THREADS = 'THREADS';
+    case TELEGRAM = 'TELEGRAM';
+}