+++ src/Components/ViewInBrowser.php
@@ -4,15 +4,14 @@ namespace Templateless\Components;
 
 use JsonSerializable;
 use Templateless\Components\ComponentId;
-
-interface Component {}
+use Templateless\Components\Component;
 
 class ViewInBrowser implements Component, JsonSerializable
 {
     private $id;
     private $text;
 
-    public function __construct($text)
+    public function __construct($text = '')
     {
         $this->id = ComponentId::VIEW_IN_BROWSER;
         $this->text = $text;