+++ src/Theme.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Templateless;
+
+enum Theme: string
+{
+    case UNSTYLED = 'UNSTYLED';
+    case SIMPLE = 'SIMPLE';
+}